receiver/requirements.txt
pyserial
websockets
paho-mqtt
